pyproject.toml
[project]
name = "testwindi"
version = "0.1.0"
description = ""
authors = [
    {name = "jonny",email = "egorsolomin6@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.115.8"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.38"
asyncpg = "^0.30.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = {extras = ["email"], version = "^2.10.6"}
passlib = "^1.7.4"

